Move opentracing implementation functions to separate file
